fixed some

+++ database/migrations/2024_11_06_210216_create_prices_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
         Schema::create('prices', function (Blueprint $table) {
             $table->id();
             $table->string('service_type');
-            $table->decimal('price', 10, 2);
+            $table->integer('price'); // Changed to integer
             $table->timestamps();
         });
     }
